Set undefs to empty strings for env vars
This means that env vars that are set with a value of undef are still
created, just with a value of an empty string - which apps then see as unset.
This makes the behaviour of this define consistent with the approach used
in govuk::app::ennvar [1].

Without this change, setting undef to an env var would mean puppet would
do nothing so any existing env vars would not be changed or removed.

[1]: https://github.com/alphagov/govuk-puppet/blob/c997c57e3363ff45127fb2e98f028ac131525cfe/modules/govuk/manifests/app/envvar.pp