Set up Content Security Policy common environment variables
This sets up environment variables for the content security policy.
These are set in govuk::deploy::config as this is a place where they can
easily be configured in a global way, whereas if they're set within
govuk::app we'll need to set them up for each app individually.

These environment variables are intended to replace hardcoded logic in
the code [1]. The expectation is that we may want groups of apps to have
different report uri values (e.g. frontend and backend) so may well set
up values for these at node level hieradata.

If apps need to customise their approach to these they can set up their
own environment variables.

Finally, worth noting that if/when we want to remove these env vars
we'll need to add an ensure absent to govuk_envvar::init. If we just
delete these from the code puppet won't do anything to them.

[1]: https://github.com/alphagov/govuk_app_config/blob/f0fa82e2d8a73449efabd2414d93264a31afdf75/lib/govuk_app_config/govuk_content_security_policy.rb#L131-L136